refactor: move parser tests to tests/
